CLI fixes

+++ src/rtlsdr_scan.py
@@ -57,7 +57,7 @@ def arguments():
     parser.add_argument("-s", "--start", help="Start frequency (MHz)",
                         type=int)
     parser.add_argument("-e", "--end", help="End frequency (MHz)", type=int)
-    parser.add_argument("-g", "--gain", help="Gain (dB)", type=int, default=0)
+    parser.add_argument("-g", "--gain", help="Gain (dB)", type=float, default=0)
     parser.add_argument("-d", "--dwell", help="Dwell time (seconds)",
                         type=float, default=0.1)
     parser.add_argument("-f", "--fft", help="FFT bins", type=int, default=1024)
